requirements.txt
fire==0.6.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
redis==5.0.8
redlock-py==1.0.8
six==1.16.0
termcolor==2.4.0
